+++ gevent/_compat.py
@@ -40,8 +40,10 @@ else:
 ## Functions
 if PY3:
     iteritems = dict.items
+    itervalues = dict.values
     xrange = range
 
 else:
     iteritems = dict.iteritems # python 3: pylint:disable=no-member
+    itervalues = dict.itervalues # python 3: pylint:disable=no-member
     xrange = __builtin__.xrange # python 2: pylint:disable=redefined-variable-type